requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
requests
resend
beautifulsoup4
openai>=1.0.0
psycopg2-binary
pydantic-settings
pytest>=7.0.0
pytest-cov>=4.0.0
